requirements.txt
asgiref==3.5.2
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.0.4
django-allauth==0.51.0
django-ckeditor==6.5.0
django-colorfield==0.7.2
django-crispy-forms==1.14.0
django-formtools==2.4
django-heroku==0.3.1
django-import-export==2.9.0
django-js-asset==2.0.0
django-multiselectfield==0.1.12
django-storages==1.13.1
folium==0.13.0
google-api-core==2.10.2
google-auth==2.14.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
gunicorn==20.1.0
idna==2.10
matplotlib==3.5.2
oauthlib==3.2.0
pandas==1.5.1
Pillow==9.2.0
plotly==5.10.0
plotly-express==0.4.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-decouple==3.6
python3-openid==3.2.0
pytz==2020.1
requests==2.28.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
tablib==3.2.1
twilio==7.12.0
urllib3==1.26.7
waitress==2.1.2
whitenoise==6.2.0
xhtml2pdf==0.2.8
xlwt==1.3.0
